Put controls on one hand
As suggested. The slides should now be flipped with a trigger. Move left for previous slide, move right for next slide. The threshold is also configurable on the User::Input Action Manager::User Input; Probably some testing is required. Also, controls are flipped on both hands so it shouldn't matter which controller you're using. Moving is ofc now only done in real space.